Fix Potential Null Pointer (#243)

+++ src/test/java/com/mitchellbosecke/pebble/StrictModeTest.java
@@ -3,7 +3,8 @@ package com.mitchellbosecke.pebble;
 import com.mitchellbosecke.pebble.error.PebbleException;
 import com.mitchellbosecke.pebble.error.RootAttributeNotFoundException;
 import com.mitchellbosecke.pebble.template.PebbleTemplate;
-import junit.framework.Assert;
+
+import org.junit.Assert;
 import org.junit.Test;
 
 import java.io.IOException;
@@ -41,7 +42,7 @@ public class StrictModeTest extends AbstractTest {
             Assert.fail("Exception " + RootAttributeNotFoundException.class.getCanonicalName() + " is expected.");
         } catch (RootAttributeNotFoundException e) {
             Assert.assertEquals(e.getFileName(), "templates/template.strictModeComplexExpression.peb");
-            Assert.assertEquals(e.getLineNumber(), 2);
+            Assert.assertEquals(e.getLineNumber(), (Integer) 2);
         }
     }
 
@@ -64,7 +65,7 @@ public class StrictModeTest extends AbstractTest {
             Assert.fail("Exception " + RootAttributeNotFoundException.class.getCanonicalName() + " is expected.");
         } catch (RootAttributeNotFoundException e) {
             Assert.assertEquals("templates/template.strictModeSimpleExpression.peb", e.getFileName());
-            Assert.assertEquals(2, e.getLineNumber());
+            Assert.assertEquals((Integer) 2, e.getLineNumber());
         }
     }
